+++ firmware/CHANGELOG.md
@@ -29,6 +29,7 @@ Release template (copy/paste this for new release):
 ### Added
  - new rusEFI console tab: Knock analyzer #6467
  - Suzuki K6A (NON VVTI Trigger) #6490
+ - explicit error if 32 bit java is used #6497
 
 ### Fixed
  - knock logic not activated until any configuration change via TS #6462